+++ server/server.go
@@ -2,10 +2,7 @@ package server
 
 import (
 	"context"
-	"database/sql"
 
-	sq "github.com/Masterminds/squirrel"
-	lru "github.com/hashicorp/golang-lru"
 	pb "github.com/markphelps/flipt/rpc"
 	"github.com/markphelps/flipt/storage"
 	"github.com/markphelps/flipt/storage/cache"
@@ -17,34 +14,22 @@ import (
 
 var _ pb.FliptServer = &Server{}
 
-// Option is a server option
-type Option func(s *Server)
-
-// WithCacheSize sets the cache size for the server
-func WithCacheSize(size int) Option {
-	return func(s *Server) {
-		s.cacheSize = size
-	}
-}
-
 // Server serves the Flipt backend
 type Server struct {
-	cacheSize int
-
 	logger logrus.FieldLogger
+	cache  cache.Cacher
+
 	storage.FlagStore
 	storage.SegmentStore
 	storage.RuleStore
 }
 
 // New creates a new Server
-func New(logger logrus.FieldLogger, db *sql.DB, opts ...Option) *Server {
+func New(logger logrus.FieldLogger, db *storage.DB, opts ...Option) *Server {
 	var (
-		builder = sq.StatementBuilder.RunWith(sq.NewStmtCacher(db))
-
-		flagStore    = storage.NewFlagStorage(logger, builder)
-		segmentStore = storage.NewSegmentStorage(logger, builder)
-		ruleStore    = storage.NewRuleStorage(logger, sq.NewStmtCacheProxy(db), builder)
+		flagStore    = storage.NewFlagStorage(logger, db)
+		segmentStore = storage.NewSegmentStorage(logger, db)
+		ruleStore    = storage.NewRuleStorage(logger, db)
 
 		s = &Server{
 			logger:       logger,
@@ -58,11 +43,9 @@ func New(logger logrus.FieldLogger, db *sql.DB, opts ...Option) *Server {
 		opt(s)
 	}
 
-	if s.cacheSize > 0 {
-		lru, _ := lru.New(s.cacheSize)
-
+	if s.cache != nil {
 		// wrap flagStore with lru cache
-		s.FlagStore = cache.NewFlagCache(logger, lru, flagStore)
+		s.FlagStore = cache.NewFlagCache(logger, s.cache, flagStore)
 	}
 
 	return s