+++ server/server_test.go
@@ -51,7 +51,6 @@ func TestErrorUnaryInterceptor(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-
 			var (
 				subject = &Server{}