+++ storage/flag.go
@@ -6,6 +6,7 @@ import (
 
 	sq "github.com/Masterminds/squirrel"
 	"github.com/gofrs/uuid"
+	"github.com/lib/pq"
 
 	proto "github.com/golang/protobuf/ptypes"
 	flipt "github.com/markphelps/flipt/rpc"
@@ -17,15 +18,18 @@ var _ FlagStore = &FlagStorage{}
 
 // FlagStorage is a SQL FlagStore
 type FlagStorage struct {
-	logger  logrus.FieldLogger
-	builder sq.StatementBuilderType
+	logger logrus.FieldLogger
+	*DB
 }
 
 // NewFlagStorage creates a FlagStorage
-func NewFlagStorage(logger logrus.FieldLogger, builder sq.StatementBuilderType) *FlagStorage {
+func NewFlagStorage(logger logrus.FieldLogger, db *DB) *FlagStorage {
 	return &FlagStorage{
-		logger:  logger.WithField("storage", "flag"),
-		builder: builder,
+		logger: logger.WithFields(logrus.Fields{
+			"storage": "flag",
+			"db":      db.dbType,
+		}),
+		DB: db,
 	}
 }
 
@@ -156,11 +160,19 @@ func (s *FlagStorage) CreateFlag(ctx context.Context, r *flipt.CreateFlagRequest
 	)
 
 	if _, err := query.ExecContext(ctx); err != nil {
-		if sqliteErr, ok := err.(sqlite3.Error); ok {
-			if sqliteErr.Code == sqlite3.ErrConstraint {
+		switch ierr := err.(type) {
+		case sqlite3.Error:
+			if ierr.Code == sqlite3.ErrConstraint {
 				return nil, ErrInvalidf("flag %q is not unique", r.Key)
 			}
+		case *pq.Error:
+			if ierr.Code.Name() == pgUniqueConstraint {
+				return nil, ErrInvalidf("flag %q is not unique", r.Key)
+			}
+		default:
+			s.logger.Infof("%+v", ierr)
 		}
+
 		return nil, err
 	}
 
@@ -231,11 +243,17 @@ func (s *FlagStorage) CreateVariant(ctx context.Context, r *flipt.CreateVariantR
 	)
 
 	if _, err := query.ExecContext(ctx); err != nil {
-		if sqliteErr, ok := err.(sqlite3.Error); ok {
-			if sqliteErr.ExtendedCode == sqlite3.ErrConstraintForeignKey {
+		switch ierr := err.(type) {
+		case sqlite3.Error:
+			if ierr.Code == sqlite3.ErrConstraint {
+				return nil, ErrNotFoundf("flag %q", r.FlagKey)
+			}
+		case *pq.Error:
+			if ierr.Code.Name() == pgForeignKeyConstraint {
 				return nil, ErrNotFoundf("flag %q", r.FlagKey)
 			}
 		}
+
 		return nil, err
 	}
 
@@ -254,7 +272,6 @@ func (s *FlagStorage) UpdateVariant(ctx context.Context, r *flipt.UpdateVariantR
 		Set("updated_at", &timestamp{proto.TimestampNow()}).
 		Where(sq.And{sq.Eq{"id": r.Id}, sq.Eq{"flag_key": r.FlagKey}}).
 		ExecContext(ctx)
-
 	if err != nil {
 		return nil, err
 	}