+++ storage/segment.go
@@ -7,6 +7,7 @@ import (
 
 	sq "github.com/Masterminds/squirrel"
 	"github.com/gofrs/uuid"
+	"github.com/lib/pq"
 
 	proto "github.com/golang/protobuf/ptypes"
 	flipt "github.com/markphelps/flipt/rpc"
@@ -18,15 +19,18 @@ var _ SegmentStore = &SegmentStorage{}
 
 // SegmentStorage is a SQL SegmentStore
 type SegmentStorage struct {
-	logger  logrus.FieldLogger
-	builder sq.StatementBuilderType
+	logger logrus.FieldLogger
+	*DB
 }
 
 // NewSegmentStorage creates a SegmentStorage
-func NewSegmentStorage(logger logrus.FieldLogger, builder sq.StatementBuilderType) *SegmentStorage {
+func NewSegmentStorage(logger logrus.FieldLogger, db *DB) *SegmentStorage {
 	return &SegmentStorage{
-		logger:  logger.WithField("storage", "segment"),
-		builder: builder,
+		logger: logger.WithFields(logrus.Fields{
+			"storage": "segment",
+			"db":      db.dbType,
+		}),
+		DB: db,
 	}
 }
 
@@ -153,11 +157,17 @@ func (s *SegmentStorage) CreateSegment(ctx context.Context, r *flipt.CreateSegme
 	)
 
 	if _, err := query.ExecContext(ctx); err != nil {
-		if sqliteErr, ok := err.(sqlite3.Error); ok {
-			if sqliteErr.Code == sqlite3.ErrConstraint {
+		switch ierr := err.(type) {
+		case sqlite3.Error:
+			if ierr.Code == sqlite3.ErrConstraint {
+				return nil, ErrInvalidf("segment %q is not unique", r.Key)
+			}
+		case *pq.Error:
+			if ierr.Code.Name() == pgUniqueConstraint {
 				return nil, ErrInvalidf("segment %q is not unique", r.Key)
 			}
 		}
+
 		return nil, err
 	}
 
@@ -252,11 +262,17 @@ func (s *SegmentStorage) CreateConstraint(ctx context.Context, r *flipt.CreateCo
 		Values(c.Id, c.SegmentKey, c.Type, c.Property, c.Operator, c.Value, &timestamp{c.CreatedAt}, &timestamp{c.UpdatedAt})
 
 	if _, err := query.ExecContext(ctx); err != nil {
-		if sqliteErr, ok := err.(sqlite3.Error); ok {
-			if sqliteErr.ExtendedCode == sqlite3.ErrConstraintForeignKey {
+		switch ierr := err.(type) {
+		case sqlite3.Error:
+			if ierr.Code == sqlite3.ErrConstraint {
+				return nil, ErrNotFoundf("segment %q", r.SegmentKey)
+			}
+		case *pq.Error:
+			if ierr.Code.Name() == pgForeignKeyConstraint {
 				return nil, ErrNotFoundf("segment %q", r.SegmentKey)
 			}
 		}
+
 		return nil, err
 	}
 
@@ -300,7 +316,6 @@ func (s *SegmentStorage) UpdateConstraint(ctx context.Context, r *flipt.UpdateCo
 		Set("updated_at", &timestamp{proto.TimestampNow()}).
 		Where(sq.And{sq.Eq{"id": r.Id}, sq.Eq{"segment_key": r.SegmentKey}}).
 		ExecContext(ctx)
-
 	if err != nil {
 		return nil, err
 	}